spectra and tici!
