+++ system/camerad/cameras/camera_qcom2.h
@@ -3,104 +3,15 @@
 #include <memory>
 #include <utility>
 
-#include "media/cam_isp_ife.h"
-
 #include "system/camerad/cameras/camera_common.h"
 #include "system/camerad/cameras/camera_util.h"
+#include "system/camerad/cameras/tici.h"
+#include "system/camerad/cameras/spectra.h"
 #include "system/camerad/sensors/sensor.h"
 #include "common/util.h"
 
-#define FRAME_BUF_COUNT 4
 
-struct CameraConfig {
-  int camera_num;
-  VisionStreamType stream_type;
-  float focal_len;  // millimeters
-  const char *publish_name;
-  cereal::FrameData::Builder (cereal::Event::Builder::*init_camera_state)();
-  bool enabled;
-  uint32_t phy;
-};
-
-const CameraConfig WIDE_ROAD_CAMERA_CONFIG = {
-  .camera_num = 0,
-  .stream_type = VISION_STREAM_WIDE_ROAD,
-  .focal_len = 1.71,
-  .publish_name = "wideRoadCameraState",
-  .init_camera_state = &cereal::Event::Builder::initWideRoadCameraState,
-  .enabled = !getenv("DISABLE_WIDE_ROAD"),
-  .phy = CAM_ISP_IFE_IN_RES_PHY_0,
-};
-
-const CameraConfig ROAD_CAMERA_CONFIG = {
-  .camera_num = 1,
-  .stream_type = VISION_STREAM_ROAD,
-  .focal_len = 8.0,
-  .publish_name = "roadCameraState",
-  .init_camera_state = &cereal::Event::Builder::initRoadCameraState,
-  .enabled = !getenv("DISABLE_ROAD"),
-  .phy = CAM_ISP_IFE_IN_RES_PHY_1,
-};
-
-const CameraConfig DRIVER_CAMERA_CONFIG = {
-  .camera_num = 2,
-  .stream_type = VISION_STREAM_DRIVER,
-  .focal_len = 1.71,
-  .publish_name = "driverCameraState",
-  .init_camera_state = &cereal::Event::Builder::initDriverCameraState,
-  .enabled = !getenv("DISABLE_DRIVER"),
-  .phy = CAM_ISP_IFE_IN_RES_PHY_2,
-};
-
-class QcomCamera {
-public:
-  QcomCamera(MultiCameraState *multi_camera_state, const CameraConfig &config);
-  ~QcomCamera();
-
-  void camera_open();
-  void camera_close();
-  int clear_req_queue();
-  int sensors_init();
-  void config_isp(int io_mem_handle, int fence, int request_id, int buf0_mem_handle, int buf0_offset);
-  void sensors_i2c(const struct i2c_random_wr_payload* dat, int len, int op_code, bool data_word);
-
-  bool openSensor();
-  void configISP();
-  void configCSIPHY();
-  void linkDevices();
-
-  // *** state ***
-
-  bool open = false;
-  bool enabled = true;
-  CameraConfig cc;
-  MultiCameraState *multi_cam_state = nullptr;
-  std::unique_ptr<const SensorInfo> sensor;
-
-  unique_fd sensor_fd;
-  unique_fd csiphy_fd;
-
-  int32_t session_handle = -1;
-  int32_t sensor_dev_handle = -1;
-  int32_t isp_dev_handle = -1;
-  int32_t csiphy_dev_handle = -1;
-
-  int32_t link_handle = -1;
-
-  int buf0_handle = 0;
-  int buf_handle[FRAME_BUF_COUNT] = {};
-  int sync_objs[FRAME_BUF_COUNT] = {};
-  uint64_t request_ids[FRAME_BUF_COUNT] = {};
-  uint64_t request_id_last = 0;
-  uint64_t frame_id_last = 0;
-  uint64_t idx_offset = 0;
-  bool skipped = true;
-
-  CameraBuf buf;
-  MemoryManager mm;
-};
-
-class CameraState : public QcomCamera {
+class CameraState : public SpectraCamera {
 public:
   std::mutex exp_lock;
